Avoid superfluous RepositoryContent records
When adding and removing the same content in a repository version,
superfluous RepositoryContent records are created (for which version_added
== version_removed == current version).  Analogously, removing and re-adding
content creates superfluous RepositoryContent records as well.

In both cases, content is shown as added and removed.

Add a normalization method that gets rid of these records.  It needs to be
called:

- Before finalization (in order to get correct added/removed sets for
  finalization)
- After finalization (as finalization may add/remove content itself)
- If add_content()/remove_content() run into a database integrity error caused
  by adding/removing the same content multiple times.

https://pulp.plan.io/issues/5707
closes #5707